+++ src/components/Pricing/PricingCalculator/Tabbed.tsx
@@ -8,17 +8,9 @@ import { allProductsData } from '../Pricing'
 import { CallToAction } from 'components/CallToAction'
 import Link from 'components/Link'
 import useProducts from 'hooks/useProducts'
-import {
-    LogSlider,
-    NonLinearSlider,
-    inverseCurve,
-    nonLinearCurve,
-    reverseNonLinearCurve,
-    sliderCurve,
-} from '../PricingSlider/Slider'
-import Checkbox from 'components/Checkbox'
-import { NumericFormat } from 'react-number-format'
+import { LogSlider, inverseCurve, sliderCurve } from '../PricingSlider/Slider'
 import { PricingTiers } from '../Plans'
+import ProductAnalyticsTab from './Tabs/ProductAnalytics'
 
 const Modal = ({ onClose, isVisible }) => {
     return (
@@ -287,159 +279,17 @@ const Addon = ({ type, name, description, plans, addons, setAddons, volume, incl
     )
 }
 
-const AnalyticsSlider = ({ marks, min, max, className = '', label, onChange, value }) => {
-    return (
-        <div className={`${className} ml-6 relative ${label ? 'pt-7' : ''}`}>
-            {label && <p className="m-0 text-sm absolute left-0 top-0">{label}</p>}
-            <NonLinearSlider
-                stepsInRange={100}
-                marks={marks}
-                min={min}
-                max={max}
-                onChange={(value) => onChange(reverseNonLinearCurve(value))}
-                value={nonLinearCurve(value)}
-            />
-        </div>
-    )
-}
-
-const SliderToggle = ({ label = '', types, activeProduct, setAnalyticsVolume, analyticsData, ...other }) => {
-    const [volume, setVolume] = useState({})
-    const [checked, setChecked] = useState(other.checked || false)
-
-    const handleCheck = () => {
-        if (checked) {
-            const volume = {}
-            types.forEach(({ type }) => {
-                volume[type] = analyticsData[type].volume
-                setAnalyticsVolume(type, 0)
-            })
-            setVolume(volume)
-        } else {
-            types.forEach(({ type }) => {
-                setAnalyticsVolume(type, volume[type] || 0)
-            })
-        }
-        setChecked(!checked)
-    }
-
-    return (
-        <div className={`mt-2 grid ${checked ? 'mb-10' : 'mb-2 grid-cols-5'}`}>
-            <div className="space-y-3 col-span-3">
-                <Checkbox className="!text-base" checked={checked} onChange={handleCheck} value={label} />
-                {checked && (
-                    <div className="space-y-8">
-                        {types.map(({ type, label }) => (
-                            <div key={type}>
-                                <div className="grid grid-cols-5">
-                                    <AnalyticsSlider
-                                        {...activeProduct.slider}
-                                        onChange={(value) => setAnalyticsVolume(type, value)}
-                                        value={analyticsData[type].volume}
-                                        className="col-span-3"
-                                        label={label}
-                                    />
-                                    <div className="text-right font-bold m-0 self-end -mb-1.5 flex justify-end">
-                                        <NumericFormat
-                                            className="bg-transparent text-right p-0 focus:ring-0 max-w-[103px] border-t-0 border-x-0 border-dashed border-border dark:border-dark"
-                                            value={analyticsData[type].volume}
-                                            thousandSeparator=","
-                                            onValueChange={({ floatValue }) => setAnalyticsVolume(type, floatValue)}
-                                        />
-                                    </div>
-                                    <p className="text-right font-bold m-0 self-end -mb-1.5">
-                                        {formatUSD(analyticsData[type].cost)}
-                                    </p>
-                                </div>
-                            </div>
-                        ))}
-                    </div>
-                )}
-            </div>
-            {!checked && (
-                <>
-                    <span className="opacity-25 text-right">--</span>
-                    <span className="opacity-25 text-right">--</span>
-                </>
-            )}
-        </div>
-    )
-}
-
-const analyticsSliders = [
-    {
-        label: 'Website analytics',
-        types: [{ type: 'websiteAnalyticsEvents' }],
-        checked: true,
-    },
-    {
-        label: 'Product analytics',
-        types: [{ type: 'productAnalyticsEvents' }],
-        checked: true,
-    },
-    {
-        label: 'Mobile app',
-        types: [
-            { type: 'mobileAppAnonymousEvents', label: 'Anonymous events' },
-            { type: 'mobileAppAuthenticatedEvents', label: 'Events from authenticated users' },
-        ],
-    },
-    {
-        label: 'LLM events',
-        types: [
-            { type: 'llmAnonymousEvents', label: 'Anonymous events' },
-            { type: 'llmAuthenticatedEvents', label: 'Events from authenticated users' },
-        ],
-    },
-    {
-        label: 'API events',
-        types: [
-            { type: 'apiAnonymousEvents', label: 'Anonymous events' },
-            { type: 'apiAuthenticatedEvents', label: 'Events from authenticated users' },
-        ],
-    },
-    {
-        label: 'Other events',
-        types: [
-            { type: 'otherAnonymousEvents', label: 'Anonymous events' },
-            { type: 'otherAuthenticatedEvents', label: 'Events from authenticated users' },
-        ],
-    },
-]
-
-const ProductAnalyticsTab = ({ activeProduct, analyticsData, setAnalyticsVolume }) => {
-    return (
-        <div className="mb-4 pr-3">
-            <div className="grid grid-cols-5 items-end mb-2">
-                <h3 className="m-0 text-base col-span-3">Event types</h3>
-                <p className="m-0 text-right opacity-70 text-sm">Events/month</p>
-                <p className="m-0 text-right opacity-70 text-sm">Subtotal</p>
-            </div>
-
-            {analyticsSliders.map((slider) => (
-                <SliderToggle
-                    key={slider.label}
-                    analyticsData={analyticsData}
-                    setAnalyticsVolume={setAnalyticsVolume}
-                    activeProduct={activeProduct}
-                    {...slider}
-                />
-            ))}
-        </div>
-    )
-}
-
 const productTabs = {
     product_analytics: ProductAnalyticsTab,
 }
 
-const TabContent = ({ activeProduct, addons, setVolume, setAddons, setAnalyticsVolume, analyticsData }) => {
+const TabContent = ({ activeProduct, addons, setVolume, setAddons, setProduct }) => {
     const { type, cost, volume, billingData, slider, costByTier } = activeProduct
 
     return (
         <>
             <div>
-                {productTabs[activeProduct.type]?.({ activeProduct, setVolume, setAnalyticsVolume, analyticsData }) ||
+                {productTabs[activeProduct.type]?.({ activeProduct, setVolume, setProduct }) ||
                     (activeProduct.name == 'A/B testing' ? (
                         <div className="bg-accent dark:bg-accent-dark border border-light dark:border-dark rounded-md px-4 py-3 mb-2 text-sm">
                             A/B testing is currently bundled with Feature flags and shares a free tier and volume
@@ -526,7 +376,7 @@ export default function Tabbed() {
     } = useStaticQuery(allProductsData)
     const platform = billingProducts.find((product) => product.type === 'platform_and_support')
     const [activeTab, setActiveTab] = useState(0)
-    const { products, setVolume, setAnalyticsVolume, analyticsData, monthlyTotal } = useProducts()
+    const { products, setVolume, setProduct, monthlyTotal } = useProducts()
     const activeProduct = products[activeTab]
     const initialProductAddons = useMemo(() => {
         const initialAddons = []
@@ -622,8 +472,7 @@ export default function Tabbed() {
                         setAddons={setProductAddons}
                         activeProduct={activeProduct}
                         setVolume={setVolume}
-                        setAnalyticsVolume={setAnalyticsVolume}
-                        analyticsData={analyticsData}
+                        setProduct={setProduct}
                     />
                 </div>
                 <div className="md:col-span-3 pt-2 pb-0 md:pt-2.5 md:pb-2 pl-4 md:pl-3 md:pr-6 border-t border-light dark:border-dark"></div>
